requirements.txt
discord.py
python-dotenv
python-dateutil
arrow